Split common.js into three smaller scripts
